I've redone how the map rendering is done. The way it works is practically identical to before. The major differences are that: 1. The render data for tiles is stored in the TileStackRenderData2D instead of the TileStack2D. 2. TileStackRenderData2D is stored in RegionRenderer2D objects. 3. RegionRenderer2D objects are created and managed by the map renderer. 4. Before, if we weren't rendering a map, we were wasting memory with references to null render data for every potential tile on a map. Now we aren't wasting that memory.
While improving the map rendering, I created a IRequest interface and a RequestQueue class in the Utils.Threading namespace. The RequestQueue is a queue that can contain objects that implement IRequest. The queue is completely thread-safe. Its purpose is to allow you to make requests from any thread and process those requests in the proper order on one thread. I've used it for processing map rendering/unrendering requests.